+++ skunkbooth/filterManager.py
