+++ skunkbooth/filters/invert.py
