+++ skunkbooth/filters/stars.py
